# docs/04-接口与数据设计/openapi/internal-api.json
{
  "components": {
    "schemas": {
...
      "ApiResponseAttachmentResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AttachmentResponse"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseObject": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseVoid": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "AttachmentResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "currentVersion": {
            "format": "int32",
            "type": "integer"
          },
          "fileName": {
            "type": "string"
          },
          "fileSha256": {
            "type": "string"
          },
          "fileSize": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "mimeType": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "ownerType": {
            "type": "string"
          },
          "scanStatus": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "DocumentTemplateRequest": {
        "properties": {
          "domainCode": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "templateCode": {
            "type": "string"
          },
          "templateName": {
            "type": "string"
          }
        },
        "required": [
          "domainCode",
          "templateCode",
          "templateName"
        ],
        "type": "object"
      },
...
      "TemplateAttachmentRequest": {
        "properties": {
          "attachmentId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "attachmentId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "内部用户认证、RBAC、动态菜单、审批任务、系统治理及阶段1主数据接口。",
    "title": "SRM 内部管理端 API",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/master-data/organizations/{id}/disable": {
      "post": {
        "operationId": "disable_7",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用组织",
        "tags": [
          "organization-controller"
        ]
      }
    },
...
    "/api/v1/messages/my": {
      "get": {
        "operationId": "myMessages",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "我的消息分页",
        "tags": [
          "message-controller"
        ]
      }
    },
...
    "/api/v1/messages/{id}/read": {
      "post": {
        "operationId": "markRead",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Mark read message",
        "tags": [
          "message-controller"
        ]
      }
    },
...
    "/api/v1/system/attachments/{id}/download": {
      "get": {
        "operationId": "download_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "byte",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Download attachment",
        "tags": [
          "attachment-controller"
        ]
      }
    },
    "/api/v1/system/attachments/{id}/replace": {
      "post": {
        "operationId": "replace",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAttachmentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Replace attachment",
        "tags": [
          "attachment-controller"
        ]
      }
    },
...
    "/api/v1/system/batch-jobs/import-template": {
      "get": {
        "operationId": "template",
        "parameters": [
          {
            "in": "query",
            "name": "objectType",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "byte",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Download import template batch job",
        "tags": [
          "batch-job-controller"
        ]
      }
    },
...
    "/api/v1/system/departments/{id}/disable": {
      "post": {
        "operationId": "disable_6",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用部门",
        "tags": [
          "department-controller"
        ]
      }
    },
...
    "/api/v1/system/dictionaries/{id}/disable": {
      "post": {
        "operationId": "disable_5",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable dictionary",
        "tags": [
          "dictionary-controller"
        ]
      }
    },
...
    "/api/v1/system/document-templates": {
      "get": {
        "operationId": "list_5",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "templateCode",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List document template",
        "tags": [
          "document-template-controller"
        ]
      },
      "post": {
        "operationId": "create_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create document template",
        "tags": [
          "document-template-controller"
        ]
      }
    },
    "/api/v1/system/document-templates/{id}": {
      "get": {
        "operationId": "get_8",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get document template",
        "tags": [
          "document-template-controller"
        ]
      }
    },
    "/api/v1/system/document-templates/{id}/attachment": {
      "put": {
        "operationId": "bindAttachment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateAttachmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Bind attachment document template",
        "tags": [
          "document-template-controller"
        ]
      }
    },
    "/api/v1/system/document-templates/{id}/disable": {
      "post": {
        "operationId": "disable_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable document template",
        "tags": [
          "document-template-controller"
        ]
      }
    },
    "/api/v1/system/document-templates/{id}/download": {
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "byte",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Download document template",
        "tags": [
          "document-template-controller"
        ]
      }
    },
    "/api/v1/system/document-templates/{id}/publish": {
      "post": {
        "operationId": "publish_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Publish document template",
        "tags": [
          "document-template-controller"
        ]
      }
    },
...
    "/api/v1/system/number-rules/{id}/disable": {
      "post": {
        "operationId": "disable_3",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Disable number rule",
        "tags": [
          "number-rule-controller"
        ]
      }
    },
...
    "/api/v1/system/positions/{id}/disable": {
      "post": {
        "operationId": "disable_2",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用岗位",
        "tags": [
          "position-controller"
        ]
      }
    },
...
    "/api/v1/system/roles/{id}/disable": {
      "post": {
        "operationId": "disable_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用角色",
        "tags": [
          "role-controller"
        ]
      }
    },
...
    "/api/v1/system/users/{id}/disable": {
      "post": {
        "operationId": "disable",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "停用用户",
        "tags": [
          "user-controller"
        ]
      }
    },
...
    "/api/v1/system/workflows/{id}/publish": {
      "post": {
        "operationId": "publish",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Publish workflow",
        "tags": [
          "workflow-controller"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "通过当前站点的同源网关访问",
      "url": "/"
    }
  ]
}